Refactor MovieList and Movie
