+++ app/components/movieList/MovieList.jsx
@@ -22,13 +22,7 @@ const MovieList = (props) => {
 
   const movies = props.movies.map(obj => {
     return (
-      <MovieCard
-        addFav={props.addFav}
-        removeFav={props.removeFav}
-        key={obj.id}
-        user={props.user ? props.user.id : null}
-        favorites={props.favorites}
-        {...obj} />
+      <MovieCard key={obj.id} {...obj} />
     );
   });
