+++ app/containers/movieListContainer/MovieListContainer.js
@@ -1,18 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { addToFavorites, removeFavorites } from '../../actions';
 import MovieList from '../../components/movieList/MovieList';
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    addFav: (movie) => {
-      dispatch(addToFavorites(movie));
-    },
-    removeFav: (movie) => {
-      dispatch(removeFavorites(movie))
-    }
-  };
-};
 
 const mapStateToProps = (state) => {
   return {
@@ -22,4 +10,4 @@ const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(MovieList)
+export default connect(mapStateToProps)(MovieList)
